@@ -1,6 +1,6 @@
 from datetime import datetime
 
 
-def log(source, message):
+def log(source: str, message: str) -> None:
     timestamp = datetime.utcnow().isoformat(timespec="seconds")
     print(f"[{timestamp}Z] [{source}] {message}")
